+++ examples/blog-mvc/src/providers/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Application Service Providers
+ *
+ * Export all providers for easy importing in bootstrap.
+ * Providers are registered in the order they are listed.
+ */
+
+export { DatabaseProvider } from './DatabaseProvider'
+export { DevEnvironmentProvider } from './DevEnvironmentProvider'
+export { MiddlewareProvider } from './MiddlewareProvider'
+export { RouteProvider } from './RouteProvider'
+export { StorageProvider } from './StorageProvider'