+++ packages/create-gravito-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-gravito-app
 
+## 1.0.1
+
+### Patch Changes
+
+- @gravito/pulse@1.0.1
+
 ## 1.0.0
 
 ### Patch Changes