+++ satellites/catalog/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gravito/satellite-catalog
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @gravito/signal@1.0.1
+
 ## 0.1.1
 
 ### Patch Changes